separate component for GeneTrack section
