+++ apps/genetics/src/sections/studyLocus/Genes/index.js
@@ -1,0 +1,1 @@
+export { default } from './StudyLocusGenes';
